Refactor latticefold::nifs::linearization::tests to use generic functions instead of macro Move linearization tests into tests folder
