+++ latticefold/src/nifs/linearization.rs
@@ -20,6 +20,7 @@ use rayon::prelude::*;
 
 pub use structs::*;
 mod structs;
+#[cfg(test)]
 mod tests;
 mod utils;
